refactor: 최소 주문 금액 객체 분리
